backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ── LinkedIn scraping ──────────────────────────────────────────────────────────
playwright==1.49.0          # pip install playwright && playwright install chromium

# ── AI / Claude ────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── HTTP client (enrichment APIs, SendGrid) ────────────────────────────────────
httpx==0.27.2

# ── Email verification (optional) ─────────────────────────────────────────────
dnspython==2.7.0            # DNS MX checks for SMTP verification

# ── Dev / testing ──────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx                       # also used by FastAPI TestClient
